+++ chicken-services/test-client-data.js
@@ -45,6 +45,9 @@ var client = {
             'SeasonName': '2020, Short Rain',
             'SeasonStart': '2020-08-01T00:00:00',
             'TotalCredit': 128000.000,
+            'TotalCreditPerCycle': {
+                '2021A': 128000.000
+            },
             'TotalRepayment_IncludingOverpayments': 1500000.0000,
             'Balance': 128000.0000,
             'CurrencyCode': 'UGX'