+++ client-enrollment/clientEnrollment.test.js
@@ -60,7 +60,7 @@ describe('clientRegistration', () => {
             roster.getClient = jest.fn().mockImplementationOnce(() => {return client ;});
             mockCursor.hasNext.mockReturnValueOnce(true);
             clientEnrollment.start(account, country, enr_lang);
-            expect(sayText).toHaveBeenCalledWith('This client is already enrolled through JiT.');
+            expect(sayText).toHaveBeenCalledWith('Thank you! You have already Enrolled with a maximum of three products.');
         });
         it('should call roster.getClinet if roster.authClient returns true', () => {
             clientEnrollment.start(account, country, enr_lang);
@@ -190,6 +190,24 @@ describe('clientRegistration', () => {
             roster.getClient = jest.fn().mockImplementationOnce(() => {return client ;});
             clientEnrollment.start(account, country,enr_lang);
             expect(clientRegistration.onContinueToEnroll).toBeCalled();
+        });
+        it('should save the  state.vars.varietyWarehouse if the variety warehouse is available', () => {
+            var {client}  = require('./test-client-with-loan');
+            client.BalanceHistory=[];
+            mockCursor.hasNext.mockReturnValueOnce(false).mockReturnValueOnce(true).mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(wareHouseRow).mockReturnValueOnce(wareHouseRow);
+            roster.getClient = jest.fn().mockImplementationOnce(() => {return client ;});
+            clientEnrollment.start(account, country,enr_lang);
+            expect(state.vars.varietyWarehouse).toBe(wareHouseRow.vars.warehouse);
+        });
+        it('should set the  state.vars.varietyWarehouse to empty if warehouse is not available', () => {
+            var {client}  = require('./test-client-with-loan');
+            client.BalanceHistory=[];
+            mockCursor.hasNext.mockReturnValueOnce(false).mockReturnValueOnce(true).mockReturnValueOnce(false);
+            mockCursor.next.mockReturnValueOnce(wareHouseRow);
+            roster.getClient = jest.fn().mockImplementationOnce(() => {return client ;});
+            clientEnrollment.start(account, country,enr_lang);
+            expect(state.vars.varietyWarehouse).toBe(' ');
         });
         it('should call not call continue to enroll if the wareHouse for that district is not found', () => {
             var {client}  = require('./test-client-with-loan');