+++ market-access/confirmation-handler/confirmationHandler.js
@@ -1,0 +1,18 @@
+var handlerName = 'confirmationHandler';
+var notifyELK = require('../../notifications/elk-notification/elkNotification');
+
+module.exports = {
+    handlerName: handlerName,
+    getHandler: function (onConfirmation){
+        return function(input){
+            notifyELK();
+            if((typeof(input) != undefined) && input != ''){
+                onConfirmation(input);
+            }
+            else{
+                onConfirmation(false);
+            }
+        };
+
+    }
+};