+++ shs/endpoints/getCode.js
@@ -10,7 +10,7 @@ module.exports = function (requestData){
         },
         {
             'unitType': 'sunking',
-            'unitSerialNumber': '23456789',
+            'unitSerialNumber': '53456562',
             'keyCode': '123 466 799',
             'keyCodeType': 'unlock'
         }