+++ shs/shs.js
@@ -8,6 +8,7 @@ var shsTypeHandler = require('./shs-type-handler/shsTypeHandler');
 var accountNumberHandler = require('./accountNumberHandler/accountNumberHandler');
 var getCodeSerialHandler = require('./get-code-serial-handler/getCodeSerialHandler');
 var registrationTypeHandler = require('./registrationTypeHandler/registrationTypeHandler');
+var replacementHandler = require('./replacementHandler/replacementHandler');
 var notifyELK = require('../notifications/elk-notification/elkNotification');
 
 
@@ -30,14 +31,17 @@ module.exports = {
         addInputHandler(getCodeSerialHandler.handlerName, getCodeSerialHandler.getHandler(onSerialValidated));
         addInputHandler(registrationTypeHandler.handlerName, registrationTypeHandler.getHandler());
         addInputHandler(gLMenuHandler.handlerName, gLMenuHandler.getHandler());
+        addInputHandler(replacementHandler.handlerName, replacementHandler.getHandler(onSerialValidated));
     },
-    start: function(account, country, lang, isGroupLeader){
+    start: function(account, country, lang, isGroupLeader,main_menu, main_menu_handler){
         notifyELK();
         state.vars.account = account;
         state.vars.country = country;
         state.vars.shsLang = lang;
         state.vars.gL = isGroupLeader;
-        state.vars.action = ' ';
+        state.vars.replacement = '';
+        state.vars.main_menu = main_menu;
+        state.vars.main_menu_handler = main_menu_handler;
         translate =  createTranslator(translations, state.vars.shsLang);
         if(isGroupLeader){
             global.sayText(translate('gl_menu',{},state.vars.shsLang));