+++ CHANGELOG.md
@@ -3,6 +3,43 @@ All notable changes to this project will be documented in this file.
 
 ## New Version
 
+
+### Added
+* [MOB-161](https://oneacrefund.atlassian.net/browse/MOB-161) Update Market Access Info
+
+### Changed
+* [SER-409](https://oneacrefund.atlassian.net/browse/SER-409) January Chicken confirmation service setup
+
+### Added
+* [MOB-123](https://oneacrefund.atlassian.net/browse/MOB-123) As a farmer receiving the repayment trial call. I need to be able to select an option to make a repayment and then make that repayment while on the call.
+
+### Changed
+* [MOB-158](https://oneacrefund.atlassian.net/browse/MOB-158) Updating the way of calculating the % repaid from Total credit to A cycle credit for Chicken confirmation Clients
+
+### v1.6.6
+### Changed
+* [SER-407](https://oneacrefund.atlassian.net/browse/SER-407) some clients are unable to top-up or enroll through JiT enrollment - changed a message being displayed
+
+### v1.6.5
+### Fixed
+* [MOB-159](https://oneacrefund.atlassian.net/browse/MOB-159) KE- GLs unable to access the Group Repayment Summary
+
+### v1.6.4
+### Changed
+* [MOB-157](https://oneacrefund.atlassian.net/browse/MOB-157) Remove the menu redirection to products from RW enrollment
+
+### v1.6.3
+### Added
+* [MOB-150](https://oneacrefund.atlassian.net/browse/MOB-150) Make JiT Top-Up and Enrollment invisible on Client Menu
+
+### v1.6.2
+### Fixed
+* [MOB-156](https://oneacrefund.atlassian.net/browse/MOB-156) Update the KE registration to check the warehouse stock
+
+## v1.6.1
+### Fixed
+* [MOB-146](https://oneacrefund.atlassian.net/browse/MOB-146) maize_recommendation was unable to automatically process an incoming SMS
+## v1.6.0
 ### Fixed
 * [MOB-109](https://oneacrefund.atlassian.net/browse/MOB-109) Fix Healthy path calculation
 ## v1.5.1